+++ app/src/main/java/com/akaiyukiusagi/quicktodo/dataLayer/room/Converters.kt
@@ -1,0 +1,19 @@
+package com.akaiyukiusagi.quicktodo.dataLayer.room
+
+import androidx.room.TypeConverter
+import java.time.LocalDateTime
+import java.time.format.DateTimeFormatter
+
+class Converters {
+    @TypeConverter
+    fun fromLocalDateTime(localDateTime: LocalDateTime?): String? {
+        return localDateTime?.format(DateTimeFormatter.ISO_LOCAL_DATE_TIME)
+    }
+
+    @TypeConverter
+    fun toLocalDateTime(localDateTimeString: String?): LocalDateTime? {
+        return localDateTimeString?.let {
+            LocalDateTime.parse(it, DateTimeFormatter.ISO_LOCAL_DATE_TIME)
+        }
+    }
+}
