--- app/src/main/java/com/akaiyukiusagi/quicktodo/data_layer/room/Converters.kt
@@ -1,19 +1,0 @@
-package com.akaiyukiusagi.quicktodo.data_layer.room
-
-import androidx.room.TypeConverter
-import java.time.LocalDateTime
-import java.time.format.DateTimeFormatter
-
-class Converters {
-    @TypeConverter
-    fun fromLocalDateTime(localDateTime: LocalDateTime?): String? {
-        return localDateTime?.format(DateTimeFormatter.ISO_LOCAL_DATE_TIME)
-    }
-
-    @TypeConverter
-    fun toLocalDateTime(localDateTimeString: String?): LocalDateTime? {
-        return localDateTimeString?.let {
-            LocalDateTime.parse(it, DateTimeFormatter.ISO_LOCAL_DATE_TIME)
-        }
-    }
-}
