+++ app/src/main/java/com/akaiyukiusagi/quicktodo/uiLayer/component/Center.kt
@@ -1,0 +1,19 @@
+package com.akaiyukiusagi.quicktodo.uiLayer.component
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+@Composable
+fun Center(
+    content: @Composable () -> Unit
+) {
+    Row(
+        horizontalArrangement = Arrangement.Center,
+        modifier = Modifier.fillMaxWidth()
+    ) {
+        content()
+    }
+}
