+++ app/src/main/java/com/akaiyukiusagi/quicktodo/uiLayer/component/PerformVibration.kt
@@ -1,0 +1,23 @@
+package com.akaiyukiusagi.quicktodo.uiLayer.component
+
+import android.content.Context
+import android.os.Build
+import android.os.VibrationEffect
+import android.os.Vibrator
+import android.os.VibratorManager
+
+/**
+ * バイブレーション実行
+ *
+ * ディレクトリ悩むところ
+ */
+fun performVibration(context: Context, durationMillis: Long) {
+    val vibrator = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
+        val vibratorManager = context.getSystemService(Context.VIBRATOR_MANAGER_SERVICE) as VibratorManager
+        vibratorManager.defaultVibrator
+    } else {
+        context.getSystemService(Context.VIBRATOR_SERVICE) as Vibrator
+    }
+
+    vibrator.vibrate(VibrationEffect.createOneShot(durationMillis, VibrationEffect.DEFAULT_AMPLITUDE))
+}
