--- app/src/main/java/com/akaiyukiusagi/quicktodo/ui_layer/component/Center.kt
@@ -1,19 +1,0 @@
-package com.akaiyukiusagi.quicktodo.ui_layer.component
-
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-
-@Composable
-fun Center(
-    content: @Composable () -> Unit
-) {
-    Row(
-        horizontalArrangement = Arrangement.Center,
-        modifier = Modifier.fillMaxWidth()
-    ) {
-        content()
-    }
-}
