+++ app/src/screenshotTest/java/com/akaiyukiusagi/quicktodo/HomeScreenshotTest.kt
@@ -5,12 +5,12 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewDynamicColors
 import androidx.compose.ui.tooling.preview.PreviewLightDark
-import com.akaiyukiusagi.quicktodo.ui_layer.component.PreviewComponent
-import com.akaiyukiusagi.quicktodo.ui_layer.screen.CardDesign
-import com.akaiyukiusagi.quicktodo.ui_layer.screen.HomeScreen
-import com.akaiyukiusagi.quicktodo.ui_layer.screen.NewTask
-import com.akaiyukiusagi.quicktodo.ui_layer.view_model.PreviewHomeViewModel
-import com.akaiyukiusagi.quicktodo.ui_layer.view_model.PreviewSettingsViewModel
+import com.akaiyukiusagi.quicktodo.uiLayer.component.PreviewComponent
+import com.akaiyukiusagi.quicktodo.uiLayer.screen.CardDesign
+import com.akaiyukiusagi.quicktodo.uiLayer.screen.HomeScreen
+import com.akaiyukiusagi.quicktodo.uiLayer.screen.NewTask
+import com.akaiyukiusagi.quicktodo.uiLayer.viewModel.PreviewHomeViewModel
+import com.akaiyukiusagi.quicktodo.uiLayer.viewModel.PreviewSettingsViewModel
 
 class HomeScreenshotTest {
